<!--
Please make sure to read the Pull Request Guidelines:
https://github.com/aws-amplify/amplify-cli/blob/dev/CONTRIBUTING.md#pull-requests
-->

#### Description of changes
When uploading studio backend files, there is a glob that scans through all files/folders in the `amplify` project directory. For Amplify projects where there are multiple Node projects, this is problematic as scanning through `node_modules` can be very resource intensive. 

Below are some benchmarks that were performed for my teams Amplify project (we currently have 40 Node lambda functions, and 6 custom resources written in JavaScript and thus having `node_modules`)

This is the benchmark _before_ the suggested change below

| Task                                                      | Time                  |
|-----------------------------------------------------------|-----------------------|
| Zipping the #current-cloud-backend                        | 1.886s                |
| Uploading the .zip file                                   | 7.487s                |
| Deleting the old S3 directory from studio-backend          | 898ms                 |
| Performing the glob on the amplify directory              | 6 minutes 41 seconds  |
| Uploading the studio-backend files to S3                  | 558ms                 |

After running the same benchmark with the proposed changes, all of the steps remained roughly the same, except for
the glob which was reduced to `1.3 seconds`. For our project, this is very significant, as this saves us almost 7 minutes per deploy.

<!--
Thank you for your Pull Request! Please provide a description above and review
the requirements below.
-->

#### Issue #13301

<!-- Also, please reference any associated PRs for documentation updates. -->

#### Description of how you validated changes

Ran amplify push both before and after and benchmarked the results for our production project. Noticed closed to a 7 minute decrease in deployment time

#### Checklist

<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [X] PR description included
- [x] `yarn test` passes
- [X] [Pull request labels](https://github.com/aws-amplify/amplify-cli/blob/dev/CONTRIBUTING.md#labels) are added


All tests are passing except for the `featureFlags.test.ts`, which should not be relevant to the changes I've made, not sure if this is a local setup issue or if tests in dev are not passing atm

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.